_printf.c

+++ _printf.c
@@ -18,7 +18,7 @@ int _printf(const char *format, ...)
 		{
 			format++;
 			if (*format == 'c')
-				count += _putchar(va_arg(args, long int));
+				count += _putchar(va_arg(args, int));
 			else if (*format == 's')
 				count += handle_string(va_arg(args, char*));
 			else if (*format == '%')
@@ -27,7 +27,7 @@ int _printf(const char *format, ...)
 				count++;
 			}
 			else if (*format == 'd' || *format == 'i')
-				count += handle_number(va_arg(args,long  int));
+				count += handle_number(va_arg(args, int));
 				
 			else
 			{
